tests/wpt/tests/workers/interfaces/DedicatedWorkerGlobalScope/postMessage/setting-postMessage.html

<!doctype html>
<title>setting postMessage</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
(async_test()).step(function() {
  var worker = new Worker('setting-postMessage.js');
  worker.onmessage = this.step_func(function(e) {
    assert_true(e.data);
    this.done();
  });
});
</script>
